cmd: add bridge, full subcommands
